refactor: wrap tooltip text and remain file for tooltips
