+++ frontend/src/components/TextToolTip/TextToolTip.style.scss
@@ -1,0 +1,3 @@
+.overlay--text-wrap {
+    white-space: pre-wrap;
+}
